MongoDB-to-BigQuery/requirements.txt
Flask==2.1.2
requests==2.27.1
ptvsd==4.3.2 # Required for debugging.
pymongo
google-cloud-bigquery
google-cloud-pubsub>=2.6.1
pymongo[srv]
dnspython
